+++ tests/OpenSimBindings/TestTypeRegistry.cpp
@@ -3,10 +3,18 @@
 #include "src/Utils/CStringView.hpp"
 
 #include <gtest/gtest.h>
+#include <OpenSim/Common/ComponentPath.h>
+#include <OpenSim/Simulation/Control/Controller.h>
+#include <OpenSim/Simulation/Model/ContactGeometry.h>
+#include <OpenSim/Simulation/Model/Force.h>
+#include <OpenSim/Simulation/Model/Model.h>
+#include <OpenSim/Simulation/Model/Probe.h>
 #include <OpenSim/Simulation/SimbodyEngine/BallJoint.h>
+#include <OpenSim/Simulation/SimbodyEngine/Constraint.h>
 #include <OpenSim/Simulation/SimbodyEngine/EllipsoidJoint.h>
 #include <OpenSim/Simulation/SimbodyEngine/FreeJoint.h>
 #include <OpenSim/Simulation/SimbodyEngine/GimbalJoint.h>
+#include <OpenSim/Simulation/SimbodyEngine/Joint.h>
 #include <OpenSim/Simulation/SimbodyEngine/PinJoint.h>
 #include <OpenSim/Simulation/SimbodyEngine/PlanarJoint.h>
 #include <OpenSim/Simulation/SimbodyEngine/ScapulothoracicJoint.h>
@@ -15,6 +23,7 @@
 
 #include <cstddef>
 #include <optional>
+#include <utility>
 #include <vector>
 
 namespace
@@ -76,4 +85,212 @@ TEST(TypeRegistry, CoordsHaveExpectedNames)
 			ASSERT_EQ(coordProp.getValue(i).getName(), tc.expectedNames[i]) << tc.name << " coordinate " << i << " has different name from expected";
 		}
 	}
+}
+
+// #298: try adding every available joint type into a blank OpenSim model to ensure
+//       that all joint types can be added without an exception/segfault
+TEST(JointRegistry, CanAddAnyJointWithoutAnExceptionOrSegfault)
+{
+	for (std::shared_ptr<OpenSim::Joint const> const& prototype : osc::JointRegistry::prototypes())
+	{
+        ASSERT_TRUE(prototype != nullptr);
+
+		// create a blank model
+        OpenSim::Model model;
+
+        // create a body
+        auto body = std::make_unique<OpenSim::Body>();
+        body->setName("onebody");
+        body->setMass(1.0);  // required
+
+        // create joint between the model's ground and the body
+        std::unique_ptr<OpenSim::Joint> joint{prototype->clone()};
+        joint->connectSocket_parent_frame(model.getGround());
+        joint->connectSocket_child_frame(*body);
+
+        // add the joint + body to the model
+        model.addJoint(joint.release());
+        model.addBody(body.release());
+
+        // initialize the model+system+state
+		//
+		// (shouldn't throw or segfault)
+        model.finalizeFromProperties();
+        model.buildSystem();
+	}
+}
+
+// #298: try adding every available contact geometry type into a blank OpenSim model
+//       to ensure that all contact geometries can be added without an exception/segfault
+TEST(ContactGeometryRegistry, CanAddAnyContactGeometryWithoutAnExceptionOrSegfault)
+{
+	for (std::shared_ptr<OpenSim::ContactGeometry const> const& prototype : osc::ContactGeometryRegistry::prototypes())
+	{
+        ASSERT_TRUE(prototype != nullptr);
+
+		// create a blank model
+        OpenSim::Model model;
+
+		// create contact geometry attached to model's ground frame
+        std::unique_ptr<OpenSim::ContactGeometry> geom{prototype->clone()};
+        geom->connectSocket_frame(model.getGround());
+
+		// add it to the model
+        model.addContactGeometry(geom.release());
+
+		// initialize the model+system+state
+        //
+        // (shouldn't throw or segfault)
+        model.finalizeFromProperties();
+        model.buildSystem();
+	}
+}
+
+// #298: try adding every available constraint to a blank OpenSim model
+//       to ensure that all of them can be added without a segfault
+//
+// (throwing is permitted, because constraints typically rely on
+//  other stuff, e.g. coordinates, existing in the model)
+TEST(ConstraintRegistry, CanAddAnyConstraintWithoutASegfault)
+{
+    for (std::shared_ptr<OpenSim::Constraint const> const& prototype : osc::ConstraintRegistry::prototypes())
+	{
+        ASSERT_TRUE(prototype != nullptr);
+
+        // create a blank model
+        OpenSim::Model model;
+
+        // default-construct the constraint
+        std::unique_ptr<OpenSim::Constraint> constraint{prototype->clone()};
+
+        // add it to the model
+        model.addConstraint(constraint.release());
+
+        // initialize the model+system+state
+		try
+		{
+            model.finalizeFromProperties();
+            model.buildSystem();
+		}
+		catch (std::exception const&)
+		{
+			// ok: it might throw because the constraint might need more information
+			//
+			// (but it definitely shouldn't segfault etc. - the error should be recoverable)
+		}
+    }
+}
+
+// #298: try adding every available force to a blank OpenSim model
+//       to ensure that all of them can be added without a segfault
+//
+// (throwing is permitted, because forces typically rely on
+//  other stuff, e.g. coordinates, existing in the model)
+TEST(ForceRegistry, CanAddAnyForceWithoutASegfault)
+{
+    for (std::shared_ptr<OpenSim::Force const> const& prototype : osc::ForceRegistry::prototypes())
+	{
+        ASSERT_TRUE(prototype != nullptr);
+
+        // create a blank model
+        OpenSim::Model model;
+
+        // default-construct the force
+        std::unique_ptr<OpenSim::Force> force{prototype->clone()};
+
+        // initialize the model+system+state
+        try
+		{
+			model.addForce(force.release());  // finalizes, so can throw
+            model.finalizeFromProperties();
+            model.buildSystem();
+        }
+        catch (std::exception const&)
+		{
+            // ok: it might throw because the constraint might need more information
+            //
+            // (but it definitely shouldn't segfault etc. - the error should be recoverable)
+        }
+    }
+}
+
+// #298: try adding every available controller to a blank OpenSim model
+//       to ensure that all of them can be added without a segfault
+TEST(ControllerRegistry, CanAddAnyControllerWithoutASegfault)
+{
+    for (std::shared_ptr<OpenSim::Controller const> const& prototype : osc::ControllerRegistry::prototypes())
+	{
+        ASSERT_TRUE(prototype != nullptr);
+
+        // create a blank model
+        OpenSim::Model model;
+
+        // default-construct the controller
+        std::unique_ptr<OpenSim::Controller> controller{prototype->clone()};
+
+		// add it to the model
+		model.addController(controller.release());
+
+        // initialize the model+system+state
+		//
+		// (doesn't seem to throw for any controller I've tested up to now)
+        model.finalizeFromProperties();
+        model.buildSystem();
+    }
+}
+
+// #298: try adding every available probe type to a blank OpenSim model
+//       to ensure that all of them can be added without a segfault
+TEST(ProbeRegistry, CanAddAnyProbeWithoutASegfault)
+{
+    for (std::shared_ptr<OpenSim::Probe const> const& prototype : osc::ProbeRegistry::prototypes())
+	{
+        ASSERT_TRUE(prototype != nullptr);
+
+        // create a blank model
+        OpenSim::Model model;
+
+        // default-construct the probe
+        std::unique_ptr<OpenSim::Probe> probe{prototype->clone()};
+
+        // add it to the model
+        model.addProbe(probe.release());
+
+        // initialize the model+system+state
+        //
+        // (doesn't seem to throw for any probe I've tested up to now)
+        model.finalizeFromProperties();
+        model.buildSystem();
+    }
+}
+
+// #298: try adding every available "ungrouped" component (i.e. a component that
+//       cannot be cleanly assigned to a known registry type) to a blank OpenSim
+//       model to ensure that all ungrouped components can be added without a
+//       segfault
+TEST(UngroupedRegistry, CanAddAnyUngroupedComponentWithoutASegfault)
+{
+    for (std::shared_ptr<OpenSim::Component const> const& prototype : osc::UngroupedRegistry::prototypes())
+    {
+        ASSERT_TRUE(prototype != nullptr);
+
+        // create a blank model
+        OpenSim::Model model;
+
+        // default-construct the component
+        std::unique_ptr<OpenSim::Component> component{prototype->clone()};
+
+        try
+        {
+            model.addComponent(component.release());
+            model.finalizeFromProperties();
+            model.buildSystem();
+        }
+        catch (std::exception const&)
+        {
+            // ok: it might throw because the component might need more information
+            //
+            // (but it definitely shouldn't segfault etc. - the error should be recoverable)
+        }
+    }
 }